l18n: add localization support
Revise the messages to make them localizable.

Note: The log messages are not marked for localization. Probably, we
want to keep log files in English for easier global troubleshooting.

Having a user run `go generate` requires a valid and up-to-date Go
environment. Rather than instructing users how to setup the environment
correctly, the `go generate` was integrated into build.bat. This reuses
the Go building environment downloaded and prepared by build.bat to
provide controllable and consistent result.

Use `make generate` on Linux.

As the zgotext.go output varies for GOARCH=386 and amd64, one had to be
chosen to provide stable output. The former is the first one to build in
build.bat.

Signed-off-by: Simon Rozman <simon@rozman.si>